+++ shiva/shiva/buffers/TensorBuffer.py
@@ -92,10 +92,10 @@ class MultiAgentTensorBuffer(ReplayBuffer):
 
     def reset(self):
         '''Resets the buffer parameters'''
-        self.obs_buffer = torch.zeros((self.max_size, self.num_agents, self.obs_dim), dtype=torch.float64, requires_grad=False)
-        self.acs_buffer = torch.zeros((self.max_size, self.num_agents, self.acs_dim), dtype=torch.float64, requires_grad=False)
-        self.rew_buffer = torch.zeros((self.max_size, self.num_agents, 1), dtype=torch.float64, requires_grad=False)
-        self.next_obs_buffer = torch.zeros((self.max_size, self.num_agents, self.obs_dim), dtype=torch.float64, requires_grad=False)
+        self.obs_buffer = torch.zeros((self.max_size, self.num_agents, self.obs_dim), dtype=torch.float32, requires_grad=False)
+        self.acs_buffer = torch.zeros((self.max_size, self.num_agents, self.acs_dim), dtype=torch.float32, requires_grad=False)
+        self.rew_buffer = torch.zeros((self.max_size, self.num_agents, 1), dtype=torch.float32, requires_grad=False)
+        self.next_obs_buffer = torch.zeros((self.max_size, self.num_agents, self.obs_dim), dtype=torch.float32, requires_grad=False)
         self.done_buffer = torch.zeros((self.max_size, self.num_agents, 1), dtype=torch.bool, requires_grad=False)
         self.current_index = 0
         self.size = 0
